[Refactoring] Use join(", ", @arguments) to build a method argument string in CodeGenerator*.pm https://bugs.webkit.org/show_bug.cgi?id=73651
Reviewed by Darin Adler.

The code in CodeGenerator*.pm to build a method argument string is really dirty
and error-prone. It is building an argument string incrementally judging whether
", " is necessary or not, like this:

    my $method = ... ? "func(" : "func(a";
    if (...) {
        $method .= $method =~ /\($/ ? "b" : ", b";
    }
    $method .= ")";

Alternatively, we can refactor the code as follows:

    my $funcName = "func";
    my @arguments;
    push(@arguments, "a") if (...);
    push(@arguments, "b") if (...);
    my $method = $funcName . "(" . join(", ", @arguments) . ")";

This patch just refactors the code, and generated .h and .cpp files should be
exactly the same as the current .h and .cpp files.

Tests: bindings/scripts/test/*

* bindings/scripts/CodeGenerator.pm:
(GetterExpression):
(SetterExpression):
* bindings/scripts/CodeGeneratorCPP.pm:
(GenerateImplementation):
* bindings/scripts/CodeGeneratorGObject.pm:
():
* bindings/scripts/CodeGeneratorJS.pm:
(GenerateImplementation):
* bindings/scripts/CodeGeneratorObjC.pm:
(GenerateImplementation):
* bindings/scripts/CodeGeneratorV8.pm:
(GenerateNormalAttrGetter):
(GenerateNormalAttrSetter):


git-svn-id: svn://svn.chromium.org/blink/trunk@101914 bbb929c8-8fbe-4397-9dbb-9b2b20218538