[Qt][WK2] Frame is not being showed in the screen when frameLoadCompleted is called https://bugs.webkit.org/show_bug.cgi?id=73850
WebKit2 has a mechanism to prevent  the composited layer tree of the outgoing page
from being torn down or updated until the incoming page has layout. However, for simple
pages, when the frame is loaded the page is not rendered in the screen.

This patch fixes this bug calling LayerTreeHostQt::scheduleLayerFlush when m_layerFlushSchedulingEnabled
is true.

Reviewed by Noam Rosenthal.

* WebProcess/WebPage/qt/LayerTreeHostQt.cpp:
(WebKit::LayerTreeHostQt::setLayerFlushSchedulingEnabled):


git-svn-id: svn://svn.chromium.org/blink/trunk@102087 bbb929c8-8fbe-4397-9dbb-9b2b20218538