DFG no longer optimises CanvasPixelArray https://bugs.webkit.org/show_bug.cgi?id=75729
Reviewed by Gavin Barraclough.

Source/JavaScriptCore:

Rename ByteArray (in its ClassInfo) to Uint8ClampedArray to match
the future name when we switch over to the new typed-array based
ImageData specification.

* runtime/JSByteArray.cpp:

Source/WebCore:

Remove the custom ClassInfo for CanvasPixelArray as that is
defeating ByteArray optimisation, and is no longer needed
anyway as it was only there to change the visible name.

* bindings/js/JSImageDataCustom.cpp:
(WebCore::toJS):

git-svn-id: svn://svn.chromium.org/blink/trunk@104324 bbb929c8-8fbe-4397-9dbb-9b2b20218538