run-api-tests: specify individual suites and tests on the command-line
Reviewed by Adam Roben.

Fixes: <http://webkit.org/b/75065> run-api-tests should be able to run individual suites and tests

The run-api-tests script will now accept a list of arguments on
the command-line that are used as prefix filters.  To run all
tests in a suite that starts with "WTF" (WTF and WTF_Vector)
use:

    $ run-api-tests WTF

To run only the tests in the "WTF" suite, not the "WTF_Vector"
suite, use:

    $ run-api-tests WTF.

* Scripts/run-api-tests: Filter tests if any prefix arguments
are passed on the command-line.

git-svn-id: svn://svn.chromium.org/blink/trunk@104343 bbb929c8-8fbe-4397-9dbb-9b2b20218538