Make it possible to type data: URLs into MiniBrowser on Windows
Fixes <http://webkit.org/b/75084> Crash when trying to navigate to a data: URL in
MiniBrowser on Windows

Reviewed by Darin Adler.

* MiniBrowser/win/BrowserWindow.cpp:
(BrowserWindow::handleMessage): Instead of checking for an "http://" prefix when deciding
whether to prepend "http://", just look for whether the typed URL contains a colon. If it
does, we assume it already has a scheme and don't modify it.


git-svn-id: svn://svn.chromium.org/blink/trunk@104621 bbb929c8-8fbe-4397-9dbb-9b2b20218538