Borders and box masks behave incorrectly with overlapping offsets https://bugs.webkit.org/show_bug.cgi?id=76137
Source/WebCore:

Reviewed by Dean Jackson.

There are some correct behaviors for -webkit-mask-box-image where
parts of the mask are missing (when the sum of the slice sizes is
greater than one dimension of the image). To render correctly
in these cases, always use a transparency layer when rendering
the mask.

Test: fast/backgrounds/mask-box-image.html

* rendering/RenderBox.cpp:
(WebCore::RenderBox::paintMaskImages):

LayoutTests:

Reviewed by Dean Jackson.

Test -webkit-mask-box-image with large slice sizes, with and
without a transform.

* fast/backgrounds/mask-box-image.html: Added.
* fast/backgrounds/resources/dot.png: Added.
* platform/mac/fast/backgrounds/mask-box-image-expected.png: Added.
* platform/mac/fast/backgrounds/mask-box-image-expected.txt: Added.

git-svn-id: svn://svn.chromium.org/blink/trunk@105091 bbb929c8-8fbe-4397-9dbb-9b2b20218538