CodeGeneratorJS.pm should overwrite the output .h/.cpp only if the bytes differ https://bugs.webkit.org/show_bug.cgi?id=76922
Reviewed by Darin Adler.

This is one of steps to stop rebuilding .h/.cpp files
generated by unchanged IDLs (bug 76836).
This patch makes a change on CodeGeneratorJS.pm so that
it overwrites the output .h/.cpp only if the bytes differ.

No tests. No change in behavior.
I manually confirmed that when I add a new attribute to Element.idl,
the time-stamps of unrelated JS*.h and JS*.cpp do not change.

* bindings/scripts/CodeGeneratorJS.pm:
(WriteData): Used UpdateFileIfChanged().


git-svn-id: svn://svn.chromium.org/blink/trunk@105809 bbb929c8-8fbe-4397-9dbb-9b2b20218538