https://bugs.webkit.org/show_bug.cgi?id=76801 Listboxes incorrectly display contents when cleared and then re-populated.
Whenever the number of items in the listbox is less than the size of listbox (number of visible items the listbox can accomodate),
we set the listbox scroll-offset to zero. The scroll-offset of the Scrollbar should also be set to 0 so that when the listbox is re-populated,
scrollbar position and the content inside the listbox are in sync.

Source/WebCore:

Patch by Joe Thomas <joethomas@motorola.com> on 2012-01-31
Reviewed by Andreas Kling.

Tests: fast/forms/listbox-clear-restore.html

* rendering/RenderListBox.cpp:
(WebCore::RenderListBox::computeLogicalHeight): Setting the scroll-offset of the Scrollbar to 0 when scrollbar is not needed.

LayoutTests:

Added test case.

Patch by Joe Thomas <joethomas@motorola.com> on 2012-01-31
Reviewed by Andreas Kling.

* fast/forms/listbox-clear-restore-expected.html: Added.
* fast/forms/listbox-clear-restore.html: Added.

git-svn-id: svn://svn.chromium.org/blink/trunk@106430 bbb929c8-8fbe-4397-9dbb-9b2b20218538