[Gtk] atk_text_get_text_at_offset() sometimes fails to provide the correct line https://bugs.webkit.org/show_bug.cgi?id=72382
Reviewed by Martin Robinson.

Source/WebCore:

Do not add unnecesary blanks at the end of a line of text.

* accessibility/gtk/WebKitAccessibleInterfaceText.cpp:
(textForRenderer): Do not just append a '\n' at the end of a line
if the linebreak for that line was already considered.

Source/WebKit/gtk:

New unit test to ensure the right line is retrieved for the first
position of each line in a multiline preformatted portion of text.

* tests/testatk.c:
(testWebkitAtkGetTextAtOffsetWithPreformattedText): New unit test.
(main): Add the new unit test.

git-svn-id: svn://svn.chromium.org/blink/trunk@106545 bbb929c8-8fbe-4397-9dbb-9b2b20218538