Source/WebCore: Web Inspector: Add the ability to jump to the source for a given frame https://bugs.webkit.org/show_bug.cgi?id=78396
Reviewed by Tim Hatcher.

* WebCore.exp.in: Add a new exported function.
* inspector/InspectorController.cpp: Add and expose the InspectorPageAgent.
(WebCore::InspectorController::InspectorController): Set m_pageAgent.
* inspector/InspectorController.h: Add m_pageAgent.
(WebCore::InspectorController::pageAgent): Return m_pageAgent.

* inspector/InspectorFrontendClientLocal.cpp:
(WebCore::InspectorFrontendClientLocal::showMainResourceForFrame): Get the inspector's frame ID
    of the frame that was passed in, and add a call to evaluate on load.
* inspector/InspectorFrontendClientLocal.h:

* inspector/front-end/InspectorFrontendAPI.js:
(InspectorFrontendAPI.showMainResourceForFrame): Add a FIXME to show the source code for the main
    resource of the given frame.

Source/WebKit2: Web Inspector: Add the ability to jump to the source for a given frame
https://bugs.webkit.org/show_bug.cgi?id=78396

Add WebKit2 API for the ability to show the source of a frame in the web inspector. This calls down
from the WK API all the way to InspectorFrontendClientLocal::showMainResourceForFrame.

Reviewed by Tim Hatcher.

* UIProcess/API/C/WKInspector.cpp:
(WKInspectorShowMainResourceForFrame): Call WebInspectorProxy::showMainResourceForFrame.
* UIProcess/API/C/WKInspector.h:

* UIProcess/WebInspectorProxy.cpp:
(WebKit::WebInspectorProxy::showMainResourceForFrame): Send a message to the web process, passing
    the frame ID of the frame we want to show the main resource of.
* UIProcess/WebInspectorProxy.h:

* WebProcess/WebPage/WebInspector.cpp:
(WebKit::WebInspector::showMainResourceForFrame): Get the WebFrame for the frame we want to show the
    main resource of from its ID. If we found a frame, call InspectorFrontendClientLocal::showMainResourceForFrame.
* WebProcess/WebPage/WebInspector.h:

* WebProcess/WebPage/WebInspector.messages.in: Add a new message.

git-svn-id: svn://svn.chromium.org/blink/trunk@107460 bbb929c8-8fbe-4397-9dbb-9b2b20218538