Node.isEqualNode() compares attributes twice. <http://webkit.org/b/78414>
Reviewed by Anders Carlsson.

Source/WebCore: 

A single pass across the attribute maps should be enough for anyone.

Added a test verifying correct behavior of Node.isEqualNode() when comparing
two elements, one of which has had attributes that were all removed,
resulting in an empty but non-null NamedNodeMap hanging off of the element.
Note that this change is not fixing a regression, I'm just adding the test
since I came close to introducing a bug here.

Test: fast/dom/isEqualNode-after-removeAttribute.html

* dom/Node.cpp:
(WebCore::Node::isEqualNode):

LayoutTests: 

* fast/dom/isEqualNode-after-removeAttribute-expected.txt: Added.
* fast/dom/isEqualNode-after-removeAttribute.html: Added.


git-svn-id: svn://svn.chromium.org/blink/trunk@107479 bbb929c8-8fbe-4397-9dbb-9b2b20218538