[Qt] Initialize QtWebContext as much as we can in its constructor. https://bugs.webkit.org/show_bug.cgi?id=79809
Reviewed by Tor Arne Vestbø.

The icon database wouldn't be initialized on the QtWebContext while in WebKitTestRunner
since it doesn't call the initialize method on it after creating it.
Remove the initialize method, move the download manager and icon database initialization
in the constructor and call initializeContextInjectedBundleClient directly in
defaultContext to prevent overriding WKTR's injected bundle client.

* UIProcess/qt/QtWebContext.cpp:
(WebKit::QtWebContext::QtWebContext):
(WebKit::QtWebContext::defaultContext):
* UIProcess/qt/QtWebContext.h:
(QtWebContext):

git-svn-id: svn://svn.chromium.org/blink/trunk@109121 bbb929c8-8fbe-4397-9dbb-9b2b20218538